Fixed navbar and footer formatting, added flexbox containers queries, fonts
