+++ front-end/src/App.js
@@ -5,10 +5,11 @@ import { AuthProvider } from './AXIOS/AuthentificationContext';
 import Home from "./Pages/Home/Home";
 import 'bootstrap/dist/css/bootstrap.css';
 import 'bootstrap/dist/js/bootstrap.js';
+import "./Shared/global.css"
 
 import Navbar from "./NavBar/Navbar";
 import Footer from "./Footer/Footer";
-import AboutUs from "./AboutUs/AboutUs";
+import AboutUs from "./Pages/AboutUs/AboutUs";
 import Products from "./Pages/Products/ProductsPage";
 import Orders from "./Pages/Orders/Orders.js";
 import OrderDetail from './Pages/Orders/OrderDetail';
@@ -25,56 +26,68 @@ function App() {
   return (
     <AuthProvider>
       <Router>
-        <Navbar />
-        <div>
-          <Routes>
-            {/* Public routes */}
-            <Route path="/" element={<Home />} />
-            <Route path="/aboutus" element={<AboutUs />} />
-            <Route path="/products" element={<Products />} />
-            <Route path="/product/:productId" element={<ProductDetailsPage />} />
-            <Route path="/cart" element={<Cart />} />
-            <Route path="/admin/login" element={<OwnerLogin />} />
+        <div className="app-container">
+          <Navbar />
+          <main className="content">
+            <Routes>
+              {/* Public routes */}
+              <Route path="/" element={<Home />} />
+              <Route path="/aboutus" element={<AboutUs />} />
+              <Route path="/products" element={<Products />} />
+              <Route path="/product/:productId" element={<ProductDetailsPage />} />
+              <Route path="/cart" element={<Cart />} />
+              <Route path="/admin/login" element={<OwnerLogin />} />
 
-            {/* Protected routes - Owner access*/}
+              {/* Protected routes - Owner access */}
+              <Route
+                path="/admin/product/update/:productId"
+                element={
+                  <ProtectedRoute>
+                    <ProductUpdatePage />
+                  </ProtectedRoute>
+                }
+              />
+              <Route
+                path="/admin/product/create"
+                element={
+                  <ProtectedRoute>
+                    <AddProductPage />
+                  </ProtectedRoute>
+                }
+              />
 
-            {/* Products */}
-            <Route path="/admin/product/update/:productId" element={
-              <ProtectedRoute>
-                <ProductUpdatePage />
-              </ProtectedRoute>
-            } />
-            <Route path="/admin/product/create" element={
-              <ProtectedRoute>
-                <AddProductPage />
-              </ProtectedRoute>
-            } />
-
-            {/* Orders */}
-            <Route path="/orders" element={
-              <ProtectedRoute>
-                <Orders />
-              </ProtectedRoute>
-            } />
-
-            <Route path="/orders/:orderId" element={
-              <ProtectedRoute>
-                <OrderDetail />
-              </ProtectedRoute>
-            } />
-            <Route path="/admin/order/update/:orderId" element={
-              <ProtectedRoute>
-                <OrderUpdatePage />
-              </ProtectedRoute>
-            } />
-        </Routes>
-
-        </div>
-        <br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/>
+              {/* Orders */}
+              <Route
+                path="/orders"
+                element={
+                  <ProtectedRoute>
+                    <Orders />
+                  </ProtectedRoute>
+                }
+              />
+              <Route
+                path="/orders/:orderId"
+                element={
+                  <ProtectedRoute>
+                    <OrderDetail />
+                  </ProtectedRoute>
+                }
+              />
+              <Route
+                path="/admin/order/update/:orderId"
+                element={
+                  <ProtectedRoute>
+                    <OrderUpdatePage />
+                  </ProtectedRoute>
+                }
+              />
+            </Routes>
+          </main>
           <Footer />
+        </div>
       </Router>
     </AuthProvider>
-  ); 
+  );
 }
 
 export default App;
