+++ front-end/src/Footer/Footer.css
@@ -1,5 +1,4 @@
 .footer-container {
-    position: fixed;
     bottom: -100%;
     left: 0;
     width: 100%;
