+++ front-end/src/Pages/Products/ProductsPage.js
@@ -63,7 +63,7 @@ const ProductPage = () => {
     }
 
     return (
-        <div className="container mt-4">
+        <div className="container mt-5">
             <br /><br />
             <h1>Our Products</h1>
             <br />
