+++ front-end/src/Shared/global.css
@@ -1,3 +1,39 @@
-.custom-mt-5 {
+/* Custom bootstrap container styling that fixes navbar spacing */
+
+.mt-7 {
     margin-top: 7rem !important;
 }
+
+.mt-6 {
+    margin-top: 5rem !important;
+}
+
+.app-container {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+  }
+  
+  .content {
+    flex: 1;
+    padding: 1rem;
+    container-type: inline-size;
+    container-name: content-box;
+  }
+  
+  footer {
+    background-color: #333;
+    color: #fff;
+    text-align: center;
+    padding: 1rem;
+  }
+  
+  @container content-box (max-height: 600px) {
+    .content {
+      padding-bottom: 4rem; 
+    }
+  }
+
+/* Global fonts */
+
+/* Background colours */
